Add files via upload
I entered google on the internet. I said afghanistan identity. I took a photo from the gkim. I uploaded it to the notaben system for trial purposes. I took a selfie from the internet. system accepted all uploads. However, the system did not confirm it as invalid. The system works fine and smoothly.