trying to create identity via unsupported document
The document is an old style Turkish driving license which is not supported by Onfido. There is no QR code or expiry date on it. So verification failed. 
Xiaomi MI MAX 3, android, 10 QKQ1.190910.002
I have successfully took the pictures of front and back sides of the ID document on Notabene but after all, the identity did not verified at all.