Add files via upload
Node ID: 608d4ed6ffbec53d
Machine Specs: Vultr VPS service is utilized. 1 CPU, 2 GB Memory, 55 GB SSD, 2000GB Bandwidth. Ubuntu 18.04 x64
Feedback: 27 data are collected during 7-day measurement process. The average RAM usage is 1.64 GB with a standard deviation of 0.12 GB. The minimum and the maximum RAM usages are 1.44 GB and 1.91 GB, respectively. The disk usage of the database are 363.87 MB and 512.05 MB at the first measurement and the last measurement, respectively. This corresponds to an average disk usage of 0.96 MB per hour. Details of the measurement data is attached.
Note: Please note that the node stopped running and restarted on its own in the evening of 5th day of data collection. This was reported to team members and team members stated that it is fine and the challenge can be submitted.