Add files via upload
Node ID: 469edf5f0f5cc499
Machine specs: 2 GB Memory, 55 GB SSD, Ubuntu 18.04 x64 (Vultr VPS).
RAM usage: RAM usage is measured 21 times during 7 days. Average RAM usage is 1.63 GB. The standard deviation is 0.11 GB. Details are provided in the attached pdf file.
Disk usage: Disk usage is measured 21 times during 7 days. The disk usage is obtained as 389.75 MB in the first measurement, whereas 555.72 MB in the last measurement. Hence, the average disk use is 1.08 MB per hour. However, larger increases in the disk usage are observed in the last 24 hours of the measurement process, that is, the average disk usage is 1.77 MB in the last 24 hours. Details are provided in the attached pdf file.
Note: The node stopped running and restarted on its own after the data collection in the evening of 4th day. This was reported to team members and team members stated that it is fine and the challenge can be submitted.